ports/nrf: add support for time.ticks_ms()
by using the systick counter with interrupt support (similar as on
stm32 port). Note that this adds some overhead since we have to
increase the milisecond timer in an irq function. But this function
may be reused later by other periodic tasks. Also it has the advantage
that we can count up with 32 bit width, whereas the systick timer
only has 24 bits. With 32 bits, it overflows every 50 days (5h for
a 24 bit counter). Possibly, to get away completly with overflows
a 64 bit counter could be used too (although this may add some
more overhead on the python binding)